[#259] Provide device registry AMQP API in hono-service-base.
Introduced base class DeviceRegistryAmqpServerBase and moved logic to there.
