+++ services/device-registry/src/main/java/org/eclipse/hono/deviceregistry/DeviceRegistryAmqpServer.java
@@ -12,82 +12,17 @@
 
 package org.eclipse.hono.deviceregistry;
 
-import io.vertx.core.json.JsonObject;
-import io.vertx.ext.healthchecks.HealthCheckHandler;
-import io.vertx.ext.healthchecks.Status;
-import org.eclipse.hono.connection.ConnectionFactory;
-import org.eclipse.hono.service.amqp.AmqpServiceBase;
-import org.eclipse.hono.service.amqp.Endpoint;
-import org.eclipse.hono.service.auth.AuthenticationConstants;
-import org.eclipse.hono.config.ServiceConfigProperties;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
 
-import java.util.Objects;
-
+import org.eclipse.hono.service.registration.DeviceRegistryAmqpServerBase;
 
 /**
- * Spring component that serves as the default implementation of Hono's device registry.
- * <p>
- * It implements Hono's <a href="https://www.eclipse.org/hono/api/Device-Registration-API/">Device Registration API</a> and
- * <a href="https://www.eclipse.org/hono/api/Credentials-API/">Credentials API</a>.
+ * Default AQMP server for Hono's example device registry.
  */
-public final class DeviceRegistryAmqpServer extends AmqpServiceBase<ServiceConfigProperties> {
-
-    private ConnectionFactory authenticationService;
+public final class DeviceRegistryAmqpServer extends DeviceRegistryAmqpServerBase {
 
     @Override
     protected String getServiceName() {
         return "Hono-DeviceRegistry";
     }
 
-    @Autowired
-    @Qualifier("amqp")
-    @Override
-    public void setConfig(final ServiceConfigProperties configuration) {
-        setSpecificConfig(configuration);
-    }
-
-    /**
-     * Sets the factory to use for creating an AMQP 1.0 connection to
-     * the Authentication service.
-     *
-     * @param factory The factory.
-     * @throws NullPointerException if factory is {@code null}.
-     */
-    @Autowired
-    @Qualifier(AuthenticationConstants.QUALIFIER_AUTHENTICATION)
-    public void setAuthenticationServiceConnectionFactory(final ConnectionFactory factory) {
-        authenticationService = Objects.requireNonNull(factory);
-    }
-
-    /**
-     * Registers this service's endpoints' readiness checks.
-     * <p>
-     * This invokes {@link Endpoint#registerReadinessChecks(HealthCheckHandler)} for all registered endpoints
-     * and it checks if the <em>Authentication Service</em> is connected.
-     *
-     * @param handler The health check handler to register the checks with.
-     */
-    @Override
-    public void registerReadinessChecks(final HealthCheckHandler handler) {
-        for (Endpoint ep : endpoints()) {
-            ep.registerReadinessChecks(handler);
-        }
-        handler.register("authentication-service-connection", status -> {
-            if (authenticationService == null) {
-                status.complete(Status.KO(new JsonObject().put("error", "no connection factory set for Authentication service")));
-            } else {
-                LOG.debug("checking connection to Authentication service");
-                authenticationService.connect(null, null, null, s -> {
-                    if (s.succeeded()) {
-                        s.result().close();
-                        status.complete(Status.OK());
-                    } else {
-                        status.complete(Status.KO(new JsonObject().put("error", "cannot connect to Authentication service")));
-                    }
-                });
-            }
-        });
-    }
 }
